fix(lazyloading): use library
Originally I tried to use a small script that would dynamically add loading="lazy" to images.
This seemed to work fine on computers but on mobile devices it would replace all images with the original URL
right away which caused blank squares as the large images took longer to download and blurred previews were never shown.

This library was supposed to be only used if a browser doesn't support lazy loading natively, but this didn't work either as it was saying on mobile phones that it was supported. Therefore I just decided to use this library for all devices as it seems more sophisticated than my 10-line script 😅